191: Address review comments for StreamPackageHeader ABI
* 191: Replace memcpy-based struct serialization with explicit fixed-width wire format
* 191: StreamPackageHeader: 9-byte little-endian layout (uint64 frame_num + uint8 flags)
* 191: Streamer uses header.serialize(); receiver uses StreamPackageHeader::deserialize()
* 191: frame_num_ changed from size_t to uint64_t to match wire format

+++ streaming/streaming_streamer/streamer.hpp
@@ -61,7 +61,7 @@ private:
   std::string connection_url_{};
   VideoStreamInfo video_stream_info_{};
   std::atomic<bool> connection_open_{false};
-  std::size_t frame_num_{0};
+  std::uint64_t frame_num_{0};
   rtc::Configuration configuration_{};
   std::shared_ptr<rtc::WebSocket> web_socket_{};
   std::shared_ptr<Peer> peer_{};
